Moved iOSIntPackage to PhotosViewController, added check optiolal for dataPhotos and replaced dataPhotos to dataPhotosSafely in the collectionView
