+++ Navigation/Profile/Photos/DataPhotos.swift
@@ -8,3 +8,9 @@
 import UIKit
 
 var dataPhotos: [UIImage?] = [UIImage(named: "1.png"), UIImage(named: "2.png"), UIImage(named: "3.png"), UIImage(named: "4.png"), UIImage(named: "5.png"), UIImage(named: "6.png"), UIImage(named: "7.png"), UIImage(named: "8.png"), UIImage(named: "9.png"), UIImage(named: "10.png"), UIImage(named: "11.png"), UIImage(named: "12.png"), UIImage(named: "13.png"), UIImage(named: "14.png"), UIImage(named: "15.png"), UIImage(named: "16.png"), UIImage(named: "17.png"), UIImage(named: "18.png"), UIImage(named: "19.png"), UIImage(named: "20.png")]
+
+var dataPhotosSafely: [UIImage] = []
+
+func checkOptional(images: [UIImage?]) {
+    images.forEach{ if $0 != nil {dataPhotosSafely.append($0!)} else {return}}
+}
